streamlit/requirements.txt
streamlit
boto3==1.33.2
python-dotenv